# docs/swagger.yaml
basePath: /
definitions:
  core.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
  core.ChatMessageRequest:
    properties:
      message:
        type: string
      order_id:
        type: string
      receiver_id:
        type: string
    required:
    - message
    - order_id
    - receiver_id
    type: object
  core.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
  core.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/core.PaginationResponse'
      status:
        type: string
      timestamp:
        type: string
    type: object
  core.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@iphoneservice.com
    name: iPhone Service API Support
    url: https://iphoneservice.com
  description: Backend API for iPhone Service Point of Sales system supporting 50
    branches across Indonesia
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: iPhone Service POS API
  version: "1.0"
paths:
...
  /chat/conversations:
    get:
      consumes:
      - application/json
      description: Get all chat conversations for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user chats
      tags:
      - chat
  /chat/messages:
    post:
      consumes:
      - application/json
      description: Send a chat message in an order conversation
      parameters:
      - description: Chat message data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/core.ChatMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send chat message
      tags:
      - chat
  /chat/messages/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a chat message as read
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark message as read
      tags:
      - chat
  /chat/orders/{orderId}/messages:
    get:
      consumes:
      - application/json
      description: Get chat messages for an order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get chat messages
      tags:
      - chat
  /chat/orders/{orderId}/read:
    put:
      consumes:
      - application/json
      description: Mark all messages in an order as read for the current user
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark order messages as read
      tags:
      - chat
  /chat/unread-count:
    get:
      consumes:
      - application/json
      description: Get the count of unread messages for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get unread message count
      tags:
      - chat
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"